Refactor Assuan code to make integration testing easier
Two facets to this change:
* Move some server-specific startup code out of agent.c into a separate
  file so that it can be used in tests.
* Compile most of the server and client code as a CMake 'object' library
  so that it can be used in tests.
