+++ src/createrepo-agent/agent.c
@@ -18,8 +18,6 @@
 #include <signal.h>
 #include <stdlib.h>
 #include <string.h>
-#include <sys/socket.h>
-#include <sys/un.h>
 
 #include <createrepo_c/createrepo_c.h>
 #include <glib.h>
@@ -29,56 +27,7 @@
 #include "createrepo-agent/command.h"
 #include "createrepo-agent/common.h"
 #include "createrepo-agent/options.h"
-
-static gpg_error_t
-try_server(const char * name)
-{
-  assuan_context_t ctx;
-  gpg_error_t rc;
-
-  rc = assuan_new(&ctx);
-  if (rc) {
-    return rc;
-  }
-
-  rc = assuan_socket_connect(ctx, name, ASSUAN_INVALID_PID, 0);
-
-  assuan_release(ctx);
-
-  return rc;
-}
-
-static assuan_fd_t
-create_server_socket(const char * name)
-{
-  struct sockaddr_un addr_un;
-  struct sockaddr * addr = (struct sockaddr *)&addr_un;
-  assuan_fd_t fd;
-  int r_redirected;
-
-  addr_un.sun_family = AF_UNIX;
-
-  if (assuan_sock_set_sockaddr_un(name, addr, &r_redirected)) {
-    return ASSUAN_INVALID_FD;
-  }
-
-  fd = assuan_sock_new(addr_un.sun_family, SOCK_STREAM, 0);
-  if (fd == ASSUAN_INVALID_FD) {
-    return ASSUAN_INVALID_FD;
-  }
-
-  if (assuan_sock_bind(fd, addr, sizeof(addr_un))) {
-    assuan_sock_close(fd);
-    return ASSUAN_INVALID_FD;
-  }
-
-  if (listen(fd, SOMAXCONN)) {
-    assuan_sock_close(fd);
-    return ASSUAN_INVALID_FD;
-  }
-
-  return fd;
-}
+#include "createrepo-agent/server.h"
 
 void
 ignore_sigpipe()
